mcp_email_server_resend/requirements.txt
mcp
resend
python-dotenv
